% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testassay.R
\name{lognormConstCVCI}
\alias{lognormConstCVCI}
\title{log-normal constant CV model}
\usage{
lognormConstCVCI(y, tau, conf.level = 0.6827)
}
\arguments{
\item{y}{Observed value}

\item{tau}{Upper limit of CI for the CV}

\item{conf.level}{Confidence level}
}
\description{
log-normal constant CV model
}

