% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testassay.R
\name{normConstCVCI}
\alias{normConstCVCI}
\title{Normal constant CV model}
\usage{
normConstCVCI(y, tau, conf.level = 0.6827, eps = .Machine$double.eps^0.25)
}
\arguments{
\item{y}{Observed value}

\item{tau}{Upper limit of CI for the CV}

\item{conf.level}{Confidence level}

\item{eps}{Limit}
}
\description{
Normal constant CV model
}

