% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testassay.R
\name{predict.assaytest}
\alias{predict.assaytest}
\title{Construct effective standard deviation intervals for observed assay values}
\usage{
\method{predict}{assaytest}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "assaytest"}

\item{newdata}{A vector of observed values}

\item{...}{additional arguments}
}
\value{
A data frame with the observed values, lower, and upper confidence
  limits
}
\description{
Computes effective standard deviation intervals for observed assay results.
These intervals have greater than 68.27% coverage.
}

