% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testassay.R
\name{print.assaytest}
\alias{print.assaytest}
\title{Print the results of an assaytest}
\usage{
\method{print}{assaytest}(x, ...)
}
\arguments{
\item{x}{An object of class "assaytest", as created by the function \link{testassay}}

\item{...}{Additional arguments, currently unused}
}
\description{
Print the results of an assaytest
}
