% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testassay.R
\name{testassay}
\alias{testassay}
\title{Hypothesis testing procedure for assay validation}
\usage{
testassay(x, m, n, q = 0.9, model = "normal", constant = "variance",
  data = NULL)
}
\arguments{
\item{x}{The vector of assay values}

\item{m}{The vector of values indicating sample membership}

\item{n}{The vector of values indicating replicate membership}

\item{q}{The confidence level, typically 0.8 or 0.9}

\item{model}{String specifying the distribution for the assay values. Valid
values are "normal" or "lognormal"}

\item{constant}{String specifying whether the variance is assumed to be
constant over the levels ("variance") or the coefficient of variation
("CV").}

\item{data}{Data frame or environment in which to look for x}
}
\value{
An object of class "assaytest", which is a list of components including a data frame of the relevant statistics calculated on x. Print, summary, predict, and plot methods are available.
}
\description{
Specify the assay results, procedure parameters, and the model assumptions,
and this function calculates the relevant statistics
}
\details{
The list has the following components
\itemize{
 \item{sumtab}{Table summarizing the experiment, includes mean values, SD or CV estimates, and upper confidence limits on those. }
 \item{Umax}{The maximum of the upper limits on the SD or CV, used in the effective SD interval calculation}
 \item{n}{The number of samples per level}
 \item{m}{The number of levels}
 \item{q}{The confidence level}
 \item{model}{The assumed model}
 \item{constant}{The parameter assumed to be constant (either SD or CV)}
 \item{alpha}{The alpha level, calculated as (1 - q)^(1 / m)}
 \item{x}{The data vector supplied by the user}
}
}

