% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testassay.R
\docType{data}
\name{gia}
\alias{gia}
\title{Growth Inhibition Assay}
\format{A data frame with six variables: \code{parasite}, \code{assay},
  \code{plate}, \code{elisa}, \code{gia}, \code{sample}, \code{meanAAgia},
  and \code{varAAgia}}
\usage{
gia
}
\description{
Data from a growth inhibition assay experiment. Samples were run repeatedly
on different assays, for two different types of parasites. gia is the value
of interest, and the meanAAgia is the sample level mean, which is the best
estimate of the "true" gia level for that sample. varAAgia is the sample level
variance.
}
\keyword{datasets}

